split corpse looting fn
